\name{myco_decision}
\alias{Myco Decision}
\title{
  Myco Decision
}
\description{
  Currently the decision function, although this still needs lots of development!
}

\usage{
respiration(C_fungal,
            N_fungal,
            NC_in_fungal_opt)
}

\arguments{
  \item{C_fungal}{
    The C in the fungi.
  }
  \item{N_fungal}{
    The N in the fungi.
  }
  \item{NC_in_fungi_opt}{
    The optimum value of the N:C ratio in the roots.
  }
}

\value{
  \item{List}{
    List of the outputs
  
    1. Demand, unitless decision
    2. N Given, kg C eq
  }
}

\author{
  Joanna Simms, joanna.x.simms@ad.helsinki.fi
}
\references{
  TODO: add a reference here
}
