\name{Myco_Growth}
\alias{myco_growth}
\docType{package}
\title{
  Mycorrhizal Growth
}
\description{
  The growth of mycorrhiza is controlled by the limitation in C or N.
}

\usage{
respiration(C_fungal,
            N_fungal,
            a,
            b)
}

\arguments{
  \item{C_fungal}{
    Carbon in the fungal biomass
  }
  \item{N_fungal}{
    Nitrogen in the fungal biomass
  }
  \item{a}{
    Factor for the C decomposition
  }
  \item{b}{
    Factor for the N decomposition
  }
}

\value{
  \item{output}{Mycorrhizal growth in terms of kg C.}
}

\author{
  Joanna Simms, joanna.x.simms@ad.helsinki.fi
}
\references{
  TODO: add a reference here
}
