\name{plant_decision}
\alias{Plant Decision}
\title{
  Plant Decision
}
\description{
  Currently the decision function, although this still needs lots of development!
}

\usage{
respiration(C_roots,
            N_roots,
            NC_in_root_opt)
}

\arguments{
  \item{C_roots}{
    The C in the roots.
  }
  \item{N_roots}{
    The N in the roots.
  }
  \item{NC_in_root_opt}{
    The optimum value of the N:C ratio in the roots.
  }
}

\value{
  \item{List}{
    List of the outputs
  
    1. Demand, unitless decision
    2. C Given, kg C
  }
}

\author{
  Joanna Simms, joanna.x.simms@ad.helsinki.fi
}
\references{
  TODO: add a reference here
}
